feat: Improve Error Handling and Validation for `status` Attribute in `Task` Class (#21)
Check the provided `new_status` against a list of `valid_statuses` to ensure it is set to a valid status. If the status is invalid, it raises `InvalidStatusError`. 

The improved error handling will allow the caller to choose how to handle an invalid status and not force a potentially confusing default status onto the Task object.
---
Changes:

- Merge `_validate_status()` method logic into `@status.setter` method:
  - The combined method performs the same check against the internal `valid_status` list as before.
  - Pass validation: set the `Task` status to the value passed.
  - Fail validation: raise`InvalidStatusError`

- Backward compatability:
  - Retain default `status` value "Pending" if not set during `init`.

- Unit tests:
  - Test setting valid statuses.
  - Test raising an exception when setting invalid statuses and edge cases.

---
closes #18